requirements.txt
aiogram
python-dotenv
motor
dnspython